+++ packages/combobox/src/ComboboxChip/index 2.ts
@@ -1,0 +1,1 @@
+export { ComboboxChip } from './ComboboxChip';