+++ packages/menu/src/Menu.story.tsx
@@ -57,7 +57,7 @@ const meta: StoryMetaType<typeof Menu> = {
               active={true}
             >
               <MenuItem active>Apple</MenuItem>
-              <MenuItem>Banana</MenuItem>
+              <MenuItem variant="destructive">Banana</MenuItem>
               <MenuItem>Carrot</MenuItem>
               <MenuItem>Dragonfruit</MenuItem>
               <MenuItem>Eggplant</MenuItem>
@@ -197,6 +197,26 @@ export const LiveExample: StoryFn<MenuProps & { size: Size }> = ({
   );
 };
 LiveExample.parameters = {
+  chromatic: {
+    disableSnapshot: true,
+  },
+};
+
+export const InitialOpen = () => {
+  return (
+    <LeafyGreenProvider>
+      <Menu
+        initialOpen
+        trigger={<Button rightGlyph={<CaretDown />}>Menu</Button>}
+      >
+        <MenuItem>Lorem</MenuItem>
+        <MenuItem>Ipsum</MenuItem>
+        <MenuItem>Adipiscing</MenuItem>
+      </Menu>
+    </LeafyGreenProvider>
+  );
+};
+InitialOpen.parameters = {
   chromatic: {
     disableSnapshot: true,
   },