+++ packages/menu/src/index.ts
@@ -1,4 +1,9 @@
 export { Menu, type MenuProps } from './Menu';
-export { FocusableMenuItem, MenuItem, type MenuItemProps } from './MenuItem';
+export {
+  FocusableMenuItem,
+  MenuItem,
+  type MenuItemProps,
+  Variant,
+} from './MenuItem';
 export { MenuSeparator } from './MenuSeparator';
 export { SubMenu, type SubMenuProps } from './SubMenu';