+++ packages/stepper/src/StepIcon/StepIcon.types.ts
@@ -1,0 +1,8 @@
+import { HTMLElementProps } from '@leafygreen-ui/lib';
+
+import { StepState } from '../Stepper';
+
+export interface StepIconProps extends HTMLElementProps<'div'> {
+  state: StepState;
+  size?: number;
+}