+++ packages/table/src/useLeafyGreenTable/TableHeaderCheckbox.tsx
@@ -1,0 +1,32 @@
+import React from 'react';
+import { Table } from '@tanstack/react-table';
+
+import Checkbox from '@leafygreen-ui/checkbox';
+import { cx } from '@leafygreen-ui/emotion';
+import { useDarkMode } from '@leafygreen-ui/leafygreen-provider';
+
+import { useRowContext } from '../Row/RowContext';
+
+import { disabledTableRowCheckStyles } from './useLeafyGreenTable.styles';
+import { LGRowData, LGTableDataType } from '.';
+
+export const TableHeaderCheckbox = ({
+  table,
+}: {
+  table: Table<LGTableDataType<LGRowData>>;
+}) => {
+  const { theme } = useDarkMode();
+  const { disabled: rowIsDisabled } = useRowContext();
+  return (
+    <Checkbox
+      className={cx({
+        [disabledTableRowCheckStyles[theme]]: rowIsDisabled,
+      })}
+      disabled={rowIsDisabled}
+      checked={table.getIsAllRowsSelected()}
+      indeterminate={table.getIsSomeRowsSelected()}
+      onChange={table.getToggleAllRowsSelectedHandler()}
+      aria-label="Select all rows"
+    />
+  );
+};