+++ packages/a11y/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @leafygreen-ui/a11y
 
+## 1.4.11
+
+### Patch Changes
+
+- Updated dependencies [dd4f3da8]
+  - @leafygreen-ui/lib@13.0.0
+
 ## 1.4.10
 
 ### Patch Changes