+++ packages/pagination/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @leafygreen-ui/pagination
 
+## 1.0.18
+
+### Patch Changes
+
+- Updated dependencies [dd4f3da8]
+- Updated dependencies [9a471879]
+- Updated dependencies [784e9d8a]
+- Updated dependencies [dd4f3da8]
+- Updated dependencies [90053e16]
+  - @leafygreen-ui/lib@13.0.0
+  - @leafygreen-ui/select@11.0.1
+  - @leafygreen-ui/icon-button@15.0.19
+  - @leafygreen-ui/typography@18.0.0
+  - @leafygreen-ui/leafygreen-provider@3.1.10
+
+## 1.0.17
+
+### Patch Changes
+
+- Updated dependencies [c5d95794]
+- Updated dependencies [324d9730]
+  - @leafygreen-ui/select@11.0.0
+  - @leafygreen-ui/typography@17.0.2
+
 ## 1.0.16
 
 ### Patch Changes