+++ packages/popover/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @leafygreen-ui/popover
 
+## 11.1.1
+
+### Patch Changes
+
+- Updated dependencies [dd4f3da8]
+  - @leafygreen-ui/lib@13.0.0
+  - @leafygreen-ui/leafygreen-provider@3.1.10
+  - @leafygreen-ui/portal@5.0.3
+
+## 11.1.0
+
+### Minor Changes
+
+- f8c77c5d: Exports a unique `contentClassName` from Popover to enable styling of Popover inner content wrapper
+- 827060f6: Allows Popover to accept a `ref`
+
 ## 11.0.18
 
 ### Patch Changes